refactor: 일정 등록 구조 변경
