+++ src/components/addCalendar/AddCalendar.styles.ts
@@ -2,6 +2,11 @@ import styled from '@emotion/styled';
 import theme from '../../styles/theme';
 import Calendar from 'react-calendar';
 
+export const Wrapper = styled.div`
+  margin: 0 auto;
+  position: relative;
+`;
+
 export const StyledCalendar = styled(Calendar)<{ selectedMethod: string }>`
   width: 336px;
