+++ src/components/addTimeTable/Table.tsx
@@ -13,13 +13,13 @@ import {
   Top,
   Wrapper,
 } from './Table.styles';
-import { TableType } from './AddTable.types';
+import { TableType } from './AddTimeTable.types';
 import Selecto from 'react-selecto';
 
 const Table = ({
   contentRef,
-  tableSelected,
-  setTableSelected,
+  selected,
+  setSelected,
   selectedMethod,
   tablePage,
   validDateChunks,
@@ -29,17 +29,17 @@ const Table = ({
   const selectoRef = useRef<any>(null);
 
   useEffect(() => {
-    if (tableSelected[tablePage]) {
-      tableSelected[tablePage].forEach((id) => {
+    if (selected[tablePage]) {
+      selected[tablePage].forEach((id) => {
         const element = document.getElementById(id);
         element?.classList.add('selected');
       });
 
       selectoRef.current.setSelectedTargets(
-        tableSelected[tablePage].map((id) => document.getElementById(id))
+        selected[tablePage].map((id) => document.getElementById(id))
       );
     }
-  }, [tablePage, tableSelected]);
+  }, [tablePage, selected]);
 
   useEffect(() => {
     selectoRef.current.setSelectedTargets([]);
@@ -60,10 +60,10 @@ const Table = ({
       document.querySelectorAll('.selected')
     ).map((node: Element) => node.id);
 
-    const newObj = { ...tableSelected };
+    const newObj = { ...selected };
     newObj[tablePage] = newArr;
 
-    setTableSelected(newObj);
+    setSelected(newObj);
   };
 
   return (
