+++ src/pages/addTime/AddTime.tsx
@@ -1,36 +1,24 @@
 import { useEffect, useRef, useState } from 'react';
-import { useNavigate, useParams } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
 import { useRecoilState } from 'recoil';
-import _ from 'lodash';
 
 import { selectedMethodState } from '../../atoms/selectedMethodAtom';
 import { isTooltipShownState } from '../../atoms/isTooltipShownAtoms';
 
-import {
-  Body,
-  Main,
-  Title,
-  TitleWrapper,
-  Wrapper,
-  CalendarWrapper,
-} from './AddTime.styles';
+import { Body, Main, Title, TitleWrapper, Wrapper } from './AddTime.styles';
 import Header from '../../components/header/Header';
 import AddToggle from '../../components/addToggle/AddToggle';
-import AddTable from '../../components/addTable/AddTable';
+import AddTimeTable from '../../components/addTimeTable/AddTimeTable';
 import AddCalendar from '../../components/addCalendar/AddCalendar';
-import BottomButton from '../../components/bottomButton/BottomButton';
 import Tooltip from '../../components/tooltip/Tooltip';
 
 import { RoomTypes } from '../../types/roomInfo';
-
-import { API } from '../../utils/API';
-import { getRange } from '../../utils/getRange';
-import { getAllTimeRange } from '../../utils/getAllTimeRange';
 import { TableSelectedTypes } from './AddTime.types';
+import { API } from '../../utils/API';
 
 const AddTime = () => {
   const { roomUUID } = useParams();
-  const navigate = useNavigate();
+  const userName = localStorage.getItem('userName') || '';
 
   const wrapperRef = useRef<HTMLDivElement>(null);
 
@@ -45,21 +33,13 @@ const AddTime = () => {
   });
 
   const { title, dates, startTime, endTime } = room;
-
-  const [timeRange, setTimeRange] = useState<number[]>([]);
-
   const isTableView = startTime !== null && endTime !== null ? true : false;
 
   const [selectedMethod, setSelectedMethod] =
     useRecoilState(selectedMethodState);
 
-  const [previousSelectedTimes, setPreviousSelectedTimes] = useState<string[]>(
-    []
-  );
   const [tableSelected, setTableSelected] = useState<TableSelectedTypes>({});
   const [calendarSelected, setCalendarSelected] = useState<string[]>([]);
-
-  const userName = localStorage.getItem('userName') || '';
 
   const [isTooltipShown, setIsTooltipShown] =
     useRecoilState<boolean>(isTooltipShownState);
@@ -72,113 +52,8 @@ const AddTime = () => {
       setRoom(data);
     };
 
-    const getPreviousSelectedTimes = async () => {
-      const { data } = await API.get(
-        `/api/room/${roomUUID}/available-time?name=${userName}`
-      );
-      setPreviousSelectedTimes(data.availableDateTimes);
-    };
-
     getRoomInfo();
-    getPreviousSelectedTimes();
   }, []);
-
-  useEffect(() => {
-    if (!isTableView) {
-      setCalendarSelected(previousSelectedTimes);
-    }
-  }, [previousSelectedTimes]);
-
-  useEffect(() => {
-    if (startTime && endTime && wrapperRef.current) {
-      if (endTime.slice(0, 2) === '00') {
-        setTimeRange(getRange(parseInt(startTime.slice(0, 2)), 24));
-      } else {
-        setTimeRange(
-          getRange(
-            parseInt(startTime.slice(0, 2)),
-            parseInt(endTime.slice(0, 2))
-          )
-        );
-      }
-
-      document.body.style.overflow = 'hidden';
-      wrapperRef.current.style.overflow = 'hidden';
-    }
-  }, [startTime, endTime]);
-
-  const goToCurrent = () => {
-    document.body.style.overflow = '';
-    (wrapperRef.current as HTMLDivElement).style.overflow = 'auto';
-
-    navigate(`/current/${roomUUID}`);
-  };
-
-  const allTimeRange = getAllTimeRange(dates, timeRange);
-
-  const putAvailableTime = async (payload: {
-    name: string;
-    hasTime: boolean;
-    availableDateTimes: string[];
-  }) => {
-    const response = await API.put(
-      `/api/room/${roomUUID}/available-time`,
-      JSON.stringify(payload)
-    );
-
-    if (response.status === 200) {
-      goToCurrent();
-    } else {
-      alert('처리 중 오류가 발생했습니다!');
-    }
-  };
-
-  const handleApplyClick = () => {
-    if (selectedMethod === 'possible') {
-      const payload = isTableView
-        ? {
-            name: userName,
-            hasTime: true,
-            availableDateTimes: Object.values(tableSelected).flat(),
-          }
-        : {
-            name: userName,
-            hasTime: false,
-            availableDateTimes: [...calendarSelected],
-          };
-
-      putAvailableTime(payload);
-    }
-
-    if (selectedMethod === 'impossible') {
-      if (isTableView) {
-        const filteredTime = _.difference(
-          allTimeRange,
-          Object.values(tableSelected).flat()
-        );
-
-        const payload = {
-          name: userName,
-          hasTime: true,
-          availableDateTimes: filteredTime,
-        };
-
-        putAvailableTime(payload);
-      } else {
-        const newDates = dates.map((date) => `${date} 00:00`);
-        const filteredTime =
-          calendarSelected && _.difference(newDates, calendarSelected);
-
-        const payload = {
-          name: userName,
-          hasTime: false,
-          availableDateTimes: filteredTime,
-        };
-
-        putAvailableTime(payload);
-      }
-    }
-  };
 
   return (
     <Wrapper ref={wrapperRef}>
@@ -197,32 +72,25 @@ const AddTime = () => {
           <Title>시간으로 선택해 주세요</Title>
         </TitleWrapper>
         <Main>
-          {isTableView ? (
-            <AddTable
+          {startTime !== null && endTime !== null ? (
+            <AddTimeTable
+              wrapperRef={wrapperRef}
               selectedMethod={selectedMethod}
+              startTime={parseInt(startTime)}
+              endTime={parseInt(endTime)}
+              selected={tableSelected}
+              setSelected={setTableSelected}
               dates={dates}
-              timeRange={timeRange}
-              previousSelectedTimes={previousSelectedTimes}
-              tableSelected={tableSelected}
-              setTableSelected={setTableSelected}
             />
           ) : (
-            <CalendarWrapper>
-              <AddCalendar
-                dates={dates}
-                selected={calendarSelected}
-                setSelected={setCalendarSelected}
-                selectedMethod={selectedMethod}
-              />
-            </CalendarWrapper>
+            <AddCalendar
+              dates={dates}
+              selected={calendarSelected}
+              setSelected={setCalendarSelected}
+              selectedMethod={selectedMethod}
+            />
           )}
         </Main>
-        <BottomButton
-          onClick={handleApplyClick}
-          navigate={goToCurrent}
-          text="등록하기"
-          isActivated={true}
-        />
       </Body>
 
       {isTableView && isTooltipShown && (
